Fix amount in constructor int -> long

+++ Core/Generator.cs
@@ -59,7 +59,7 @@ namespace Core {
             return b;
         }
 
-        public Block CreateDistributionBlock(int amount) {
+        public Block CreateDistributionBlock(long amount) {
             List<Transaction> transactions = new List<Transaction>();
             foreach(ECDsaKey key in keys) {
                 transactions.Add(CreateDistributionTransaction(key, amount));
@@ -111,7 +111,7 @@ namespace Core {
             return t;
 		}
 
-        public Transaction CreateDistributionTransaction(ECDsaKey output, int amount) {
+        public Transaction CreateDistributionTransaction(ECDsaKey output, long amount) {
             byte[] merkleHash;
 
             if (lastSignatureDictionary.ContainsKey(genesisKey)) {
